pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi[all]>=0.125.0",
    "httpx>=0.28.1",
    "markdownify>=1.2.2",
    "openai>=2.13.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
